TIP-428: move writers
