+++ features/Context/fixtures/connector_deprecated_batch_jobs.yml
@@ -15,13 +15,13 @@ connector:
 #                    services:
 #                        reader:    pim_base_connector.reader.file.csv_product
 #                        processor: pim_connector.processor.denormalization.product.flat
-#                        writer:    pim_connector.writer.doctrine.product
+#                        writer:    pim_connector.writer.database.product
 #                import_associations:
 #                    title:     pim_base_connector.jobs.csv_product_import.import_associations.title
 #                    services:
 #                        reader:    pim_base_connector.reader.file.csv_association
 #                        processor: pim_connector.processor.denormalization.product_association.flat
-#                        writer:    pim_connector.writer.doctrine.product_association
+#                        writer:    pim_connector.writer.database.product_association
 #        csv_attribute_import:
 #            title: pim_base_connector.jobs.csv_attribute_import.title
 #            type:  import
@@ -36,7 +36,7 @@ connector:
 #                    services:
 #                        reader:    pim_base_connector.reader.file.csv_attribute
 #                        processor: pim_connector.processor.denormalization.attribute.flat
-#                        writer:    pim_connector.writer.doctrine.attribute
+#                        writer:    pim_connector.writer.database.attribute
 #        csv_attribute_option_import:
 #            title: pim_base_connector.jobs.csv_attribute_option_import.title
 #            type:  import
@@ -51,7 +51,7 @@ connector:
 #                    services:
 #                        reader:    pim_base_connector.reader.file.csv_attribute_option
 #                        processor: pim_connector.processor.denormalization.attribute_option.flat
-#                        writer:    pim_connector.writer.doctrine.attribute_option
+#                        writer:    pim_connector.writer.database.attribute_option
 #        csv_variant_group_import:
 #            title: pim_base_connector.jobs.csv_variant_group_import.title
 #            type:  import
@@ -66,7 +66,7 @@ connector:
 #                    services:
 #                        reader:    pim_base_connector.reader.file.csv_variant_group
 #                        processor: pim_connector.processor.denormalization.variant_group.flat
-#                        writer:    pim_connector.writer.doctrine.variant_group
+#                        writer:    pim_connector.writer.database.variant_group
 #                    parameters:
 #                        batchSize: 1
 #        csv_category_import:
@@ -83,7 +83,7 @@ connector:
 #                    services:
 #                        reader:    pim_connector.reader.file.csv_category
 #                        processor: pim_connector.processor.denormalization.category.flat
-#                        writer:    pim_connector.writer.doctrine.category
+#                        writer:    pim_connector.writer.database.category
 #                    parameters:
 #                        batchSize: 1
 #        csv_group_import:
@@ -100,7 +100,7 @@ connector:
 #                    services:
 #                        reader:    pim_base_connector.reader.file.csv_group
 #                        processor: pim_connector.processor.denormalization.group.flat
-#                        writer:    pim_connector.writer.doctrine.group
+#                        writer:    pim_connector.writer.database.group
 #        csv_association_type_import:
 #            title: pim_base_connector.jobs.csv_association_type_import.title
 #            type: import
@@ -115,4 +115,4 @@ connector:
 #                    services:
 #                        reader:    pim_base_connector.reader.file.csv_association_type
 #                        processor: pim_connector.processor.denormalization.association_type.flat
-#                        writer:    pim_connector.writer.doctrine.association_type
+#                        writer:    pim_connector.writer.database.association_type
