+++ src/Pim/Bundle/ConnectorBundle/Resources/config/batch_jobs/xlsx_connector.yml
@@ -12,12 +12,12 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.xlsx_product
                         processor: pim_connector.processor.denormalization.product.flat
-                        writer:    pim_connector.writer.doctrine.product
+                        writer:    pim_connector.writer.database.product
                 import_associations:
                     services:
                         reader:    pim_connector.reader.file.xlsx_association_type
                         processor: pim_connector.processor.denormalization.product_association.flat
-                        writer:    pim_connector.writer.doctrine.product_association
+                        writer:    pim_connector.writer.database.product_association
                     parameters:
                         batch_size: 1
         xlsx_category_import:
@@ -31,7 +31,7 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.xlsx_category
                         processor: pim_connector.processor.denormalization.category.flat
-                        writer:    pim_connector.writer.doctrine.category
+                        writer:    pim_connector.writer.database.category
                     parameters:
                         batchSize: 1
         xlsx_attribute_import:
@@ -45,7 +45,7 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.xlsx_attribute
                         processor: pim_connector.processor.denormalization.attribute.flat
-                        writer:    pim_connector.writer.doctrine.attribute
+                        writer:    pim_connector.writer.database.attribute
         xlsx_attribute_option_import:
             type:  import
             steps:
@@ -57,7 +57,7 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.xlsx_attribute_option
                         processor: pim_connector.processor.denormalization.attribute_option.flat
-                        writer:    pim_connector.writer.doctrine.attribute_option
+                        writer:    pim_connector.writer.database.attribute_option
         xlsx_association_type_import:
             type: import
             steps:
@@ -69,7 +69,7 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.xlsx_association_type
                         processor: pim_connector.processor.denormalization.association_type.flat
-                        writer:    pim_connector.writer.doctrine.association_type
+                        writer:    pim_connector.writer.database.association_type
         xlsx_family_import:
             type: import
             steps:
@@ -81,7 +81,7 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.xlsx_family
                         processor: pim_connector.processor.denormalization.family.flat
-                        writer:    pim_connector.writer.doctrine.family
+                        writer:    pim_connector.writer.database.family
         xlsx_product_export:
              type:  export
              steps:
@@ -123,7 +123,7 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.xlsx_group
                         processor: pim_connector.processor.denormalization.group.flat
-                        writer:    pim_connector.writer.doctrine.group
+                        writer:    pim_connector.writer.database.group
         xlsx_variant_group_import:
             type:  import
             steps:
@@ -135,7 +135,7 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.xlsx_variant_group
                         processor: pim_connector.processor.denormalization.variant_group.flat
-                        writer:    pim_connector.writer.doctrine.variant_group
+                        writer:    pim_connector.writer.database.variant_group
                     parameters:
                         batchSize: 1
         xlsx_family_export:
