+++ src/Pim/Bundle/EnrichBundle/Resources/config/batch_jobs.yml
@@ -8,7 +8,7 @@ connector:
                     services:
                         reader:    pim_enrich.connector.reader.mass_edit.product
                         processor: pim_enrich.connector.processor.mass_edit.product.update_value
-                        writer:    pim_connector.writer.doctrine.product
+                        writer:    pim_connector.writer.database.product
         add_product_value:
             type:  mass_edit
             steps:
@@ -16,7 +16,7 @@ connector:
                     services:
                         reader:    pim_enrich.connector.reader.mass_edit.product
                         processor: pim_enrich.connector.processor.mass_edit.product.add_value
-                        writer:    pim_connector.writer.doctrine.product
+                        writer:    pim_connector.writer.database.product
         edit_common_attributes:
             type:  mass_edit
             steps:
@@ -24,7 +24,7 @@ connector:
                     services:
                         reader:    pim_enrich.connector.reader.mass_edit.product
                         processor: pim_enrich.connector.processor.mass_edit.product.edit_common_attributes
-                        writer:    pim_connector.writer.doctrine.product
+                        writer:    pim_connector.writer.database.product
                 clean:
                     class: "%pim_enrich.step.mass_edit.step.class%"
                     services:
@@ -36,7 +36,7 @@ connector:
                     services:
                         reader:    pim_enrich.connector.reader.mass_edit.family
                         processor: pim_enrich.connector.processor.mass_edit.family.set_requirements
-                        writer:    pim_connector.writer.doctrine.family
+                        writer:    pim_connector.writer.database.family
         csv_product_quick_export:
             type: quick_export
             steps:
@@ -84,4 +84,4 @@ connector:
                     services:
                         reader:    pim_enrich.connector.reader.mass_edit.variant_group_product
                         processor: pim_enrich.connector.processor.mass_edit.product.add_to_variant_group
-                        writer:    pim_connector.writer.doctrine.product
+                        writer:    pim_connector.writer.database.product
