+++ src/Pim/Bundle/InstallerBundle/Resources/config/batch_jobs.yml
@@ -12,7 +12,7 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.csv_category
                         processor: pim_connector.processor.denormalization.category.flat
-                        writer:    pim_connector.writer.doctrine.category
+                        writer:    pim_connector.writer.database.category
                     parameters:
                         batchSize: 1
 
@@ -27,7 +27,7 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.csv_attribute_group
                         processor: pim_connector.processor.denormalization.attribute_group.flat
-                        writer:    pim_connector.writer.doctrine.attribute_group
+                        writer:    pim_connector.writer.database.attribute_group
 
         fixtures_channel_csv:
             type:  fixtures
@@ -40,7 +40,7 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.csv_channel
                         processor: pim_connector.processor.denormalization.channel.flat
-                        writer:    pim_connector.writer.doctrine.channel
+                        writer:    pim_connector.writer.database.channel
 
         fixtures_locale_csv:
             type:  fixtures
@@ -53,7 +53,7 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.csv_locale
                         processor: pim_connector.processor.denormalization.locale.flat
-                        writer:    pim_connector.writer.doctrine.locale
+                        writer:    pim_connector.writer.database.locale
 
         fixtures_currency_csv:
             type:  fixtures
@@ -66,7 +66,7 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.csv_currency
                         processor: pim_connector.processor.denormalization.currency.flat
-                        writer:    pim_connector.writer.doctrine.currency
+                        writer:    pim_connector.writer.database.currency
 
         fixtures_group_type_csv:
             type:  fixtures
@@ -79,7 +79,7 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.csv_group_type
                         processor: pim_connector.processor.denormalization.group_type.flat
-                        writer:    pim_connector.writer.doctrine.group_type
+                        writer:    pim_connector.writer.database.group_type
 
         fixtures_association_type_csv:
             type:  fixtures
@@ -92,7 +92,7 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.csv
                         processor: pim_connector.processor.denormalization.association_type.flat
-                        writer:    pim_connector.writer.doctrine.association_type
+                        writer:    pim_connector.writer.database.association_type
 
         fixtures_attribute_csv:
             type:  fixtures
@@ -105,7 +105,7 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.csv_attribute
                         processor: pim_connector.processor.denormalization.attribute.flat
-                        writer:    pim_connector.writer.doctrine.attribute
+                        writer:    pim_connector.writer.database.attribute
 
         fixtures_attribute_options_csv:
             type:  fixtures
@@ -118,7 +118,7 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.csv_attribute_option
                         processor: pim_connector.processor.denormalization.attribute_option.flat
-                        writer:    pim_connector.writer.doctrine.attribute_option
+                        writer:    pim_connector.writer.database.attribute_option
 
         fixtures_family_csv:
             type:  fixtures
@@ -131,7 +131,7 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.csv
                         processor: pim_connector.processor.denormalization.family.flat
-                        writer:    pim_connector.writer.doctrine.family
+                        writer:    pim_connector.writer.database.family
 
         fixtures_variant_group_csv:
             type:  fixtures
@@ -144,7 +144,7 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.csv_variant_group
                         processor: pim_connector.processor.denormalization.variant_group.flat
-                        writer:    pim_connector.writer.doctrine.variant_group
+                        writer:    pim_connector.writer.database.variant_group
 
         fixtures_product_csv:
             type:  fixtures
@@ -157,12 +157,12 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.csv_product
                         processor: pim_connector.processor.denormalization.product.flat
-                        writer:    pim_connector.writer.doctrine.product
+                        writer:    pim_connector.writer.database.product
                 import_associations:
                     services:
                         reader:    pim_connector.reader.file.csv_association
                         processor: pim_connector.processor.denormalization.product_association.flat
-                        writer:    pim_connector.writer.doctrine.product_association
+                        writer:    pim_connector.writer.database.product_association
                     parameters:
                         batch_size: 1
 
@@ -177,7 +177,7 @@ connector:
                     services:
                         reader:    pim_base_connector.reader.file.yaml
                         processor: pim_connector.processor.denormalization.job_instance.flat
-                        writer:    pim_connector.writer.doctrine.job_instance
+                        writer:    pim_connector.writer.database.job_instance
 
         fixtures_user_csv:
             type:  fixtures
@@ -190,7 +190,7 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.csv
                         processor: pim_connector.processor.denormalization.user.flat
-                        writer:    pim_connector.writer.doctrine.user
+                        writer:    pim_connector.writer.database.user
 
         fixtures_user_role_csv:
             type:  fixtures
@@ -203,7 +203,7 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.csv
                         processor: pim_connector.processor.denormalization.user_role.flat
-                        writer:    pim_connector.writer.doctrine.user_role
+                        writer:    pim_connector.writer.database.user_role
 
         fixtures_user_group_csv:
             type:  fixtures
@@ -216,7 +216,7 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.csv
                         processor: pim_connector.processor.denormalization.user_group.flat
-                        writer:    pim_connector.writer.doctrine.user_group
+                        writer:    pim_connector.writer.database.user_group
 
         fixtures_group_csv:
             type:  fixtures
@@ -229,4 +229,4 @@ connector:
                     services:
                         reader:    pim_connector.reader.file.csv_group
                         processor: pim_connector.processor.denormalization.group.flat
-                        writer:    pim_connector.writer.doctrine.group
+                        writer:    pim_connector.writer.database.group
