+++ src/Pim/Component/Connector/spec/Writer/Database/BaseWriterSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spec\Pim\Component\Connector\Writer\Doctrine;
+namespace spec\Pim\Component\Connector\Writer\Database;
 
 use Akeneo\Component\Batch\Model\StepExecution;
 use Akeneo\Component\StorageUtils\Detacher\BulkObjectDetacherInterface;
